+++ src/package.json
@@ -66,6 +66,8 @@
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@mui/base": "^5.0.0-beta.23",
-    "@mui/material": "^5.14.17"
+    "@mui/icons-material": "^5.14.18",
+    "@mui/material": "^5.14.18",
+    "re-resizable": "^6.9.11"
   }
 }
