+++ src/resources/js/components/Icons/index.jsx
@@ -1,2 +1,7 @@
 import DropdownIcon from "./svg/DownArrow";
 export { DropdownIcon };
+
+import HomeOutlinedIcon from "@mui/icons-material/HomeOutlined";
+export const HomeIcon = (props) => {
+  return <HomeOutlinedIcon {...props} />;
+};
