+++ src/resources/js/pages/Top.jsx
@@ -17,14 +17,14 @@ export default function Top({ auth, laravelVersion, phpVersion }) {
           ) : (
             <>
               <Link
-                href={route("login")}
+                href={route("login.director")}
                 className="font-semibold text-gray-600 hover:text-gray-900 dark:text-gray-400 dark:hover:text-white focus:outline focus:outline-2 focus:rounded-sm focus:outline-red-500"
               >
                 Log in
               </Link>
 
               <Link
-                href={route("register")}
+                href={route("register.director")}
                 className="ms-4 font-semibold text-gray-600 hover:text-gray-900 dark:text-gray-400 dark:hover:text-white focus:outline focus:outline-2 focus:rounded-sm focus:outline-red-500"
               >
                 Register
